scanner: Move ScanCode classes to their own package
This groups the rather complex ScanCode implemenation and puts the
generically named parsing functions into a namespace to avoid clashes.
As a result, "ScanCode" can be removed from existing function names.
