+++ scanner/src/funTest/kotlin/scanners/scancode/ScanCodeScannerFunTest.kt
@@ -17,8 +17,9 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.scanner.scanners
+package org.ossreviewtoolkit.scanner.scanners.scancode
 
+import org.ossreviewtoolkit.scanner.scanners.AbstractScannerFunTest
 import org.ossreviewtoolkit.spdx.toSpdx
 import org.ossreviewtoolkit.utils.test.ExpensiveTag
 import org.ossreviewtoolkit.utils.test.ScanCodeTag
