+++ scanner/src/main/kotlin/storages/ClearlyDefinedStorage.kt
@@ -41,8 +41,8 @@ import org.ossreviewtoolkit.model.utils.toClearlyDefinedCoordinates
 import org.ossreviewtoolkit.model.utils.toClearlyDefinedSourceLocation
 import org.ossreviewtoolkit.scanner.ScanResultsStorage
 import org.ossreviewtoolkit.scanner.ScannerCriteria
-import org.ossreviewtoolkit.scanner.scanners.generateScannerDetails
-import org.ossreviewtoolkit.scanner.scanners.generateSummary
+import org.ossreviewtoolkit.scanner.scanners.scancode.generateScannerDetails
+import org.ossreviewtoolkit.scanner.scanners.scancode.generateSummary
 import org.ossreviewtoolkit.utils.collectMessagesAsString
 import org.ossreviewtoolkit.utils.log
 import org.ossreviewtoolkit.utils.showStackTrace
