+++ scanner/src/test/kotlin/scanners/scancode/ScanCodeResultParserTest.kt
@@ -18,7 +18,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.scanner.scanners
+package org.ossreviewtoolkit.scanner.scanners.scancode
 
 import com.fasterxml.jackson.databind.node.ArrayNode
 import com.fasterxml.jackson.databind.node.ObjectNode
@@ -42,7 +42,7 @@ class ScanCodeResultParserTest : WordSpec({
     "ScanCode 2 results" should {
         "be correctly summarized" {
             val resultFile = File("src/test/assets/mime-types-2.1.18_scancode-2.9.7.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -57,7 +57,7 @@ class ScanCodeResultParserTest : WordSpec({
     "ScanCode 3 results" should {
         "be correctly summarized" {
             val resultFile = File("src/test/assets/mime-types-2.1.18_scancode-3.0.2.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -73,7 +73,7 @@ class ScanCodeResultParserTest : WordSpec({
         "properly summarize the license findings for ScanCode 2.2.1" {
             // TODO: minimize this test case.
             val resultFile = File("src/test/assets/esprima-2.7.3_scancode-2.2.1.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -455,7 +455,7 @@ class ScanCodeResultParserTest : WordSpec({
         "properly summarize the copyright findings for ScanCode 2.2.1" {
             // TODO: minimize this test case
             val resultFile = File("src/test/assets/esprima-2.7.3_scancode-2.2.1.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -642,7 +642,7 @@ class ScanCodeResultParserTest : WordSpec({
 
         "properly summarize license findings for ScanCode 2.9.7" {
             val resultFile = File("src/test/assets/aws-java-sdk-core-1.11.160_scancode-2.9.7.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             val actualFindings =
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
@@ -669,7 +669,7 @@ class ScanCodeResultParserTest : WordSpec({
 
         "properly summarize copyright findings for ScanCode 2.9.7" {
             val resultFile = File("src/test/assets/aws-java-sdk-core-1.11.160_scancode-2.9.7.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             val actualFindings =
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
@@ -700,7 +700,7 @@ class ScanCodeResultParserTest : WordSpec({
 
     "generateDetails" should {
         "parse a ScanCode 2.9.x result file" {
-            val result = parseScanCodeResult(File("src/test/assets/mime-types-2.1.18_scancode-2.9.7.json"))
+            val result = parseResultsFile(File("src/test/assets/mime-types-2.1.18_scancode-2.9.7.json"))
 
             val details = generateScannerDetails(result)
             details.name shouldBe ScanCode.SCANNER_NAME
@@ -713,7 +713,7 @@ class ScanCodeResultParserTest : WordSpec({
         }
 
         "parse a ScanCode 3.x result file" {
-            val result = parseScanCodeResult(File("src/test/assets/mime-types-2.1.18_scancode-3.0.2.json"))
+            val result = parseResultsFile(File("src/test/assets/mime-types-2.1.18_scancode-3.0.2.json"))
 
             val details = generateScannerDetails(result)
             details.name shouldBe ScanCode.SCANNER_NAME
@@ -723,7 +723,7 @@ class ScanCodeResultParserTest : WordSpec({
         }
 
         "handle missing option properties gracefully" {
-            val result = parseScanCodeResult(File("src/test/assets/mime-types-2.1.18_scancode-3.0.2.json"))
+            val result = parseResultsFile(File("src/test/assets/mime-types-2.1.18_scancode-3.0.2.json"))
             val headers = result["headers"] as ArrayNode
             val headerObj = headers[0] as ObjectNode
             headerObj.remove("options")
@@ -734,7 +734,7 @@ class ScanCodeResultParserTest : WordSpec({
 
         "handle missing scanner version property gracefully" {
             val result =
-                parseScanCodeResult(File("src/test/assets/mime-types-2.1.18_scancode-2.9.7.json")) as ObjectNode
+                parseResultsFile(File("src/test/assets/mime-types-2.1.18_scancode-2.9.7.json")) as ObjectNode
             result.remove("scancode_version")
 
             val details = generateScannerDetails(result)
@@ -745,7 +745,7 @@ class ScanCodeResultParserTest : WordSpec({
     "mapTimeoutErrors()" should {
         "return true for scan results with only timeout errors" {
             val resultFile = File("src/test/assets/esprima-2.7.3_scancode-2.2.1.post277.4d68f9377.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -786,7 +786,7 @@ class ScanCodeResultParserTest : WordSpec({
 
         "return false for scan results without errors" {
             val resultFile = File("src/test/assets/esprima-2.7.3_scancode-2.2.1.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -800,7 +800,7 @@ class ScanCodeResultParserTest : WordSpec({
     "mapUnknownErrors()" should {
         "return true for scan results with only memory errors" {
             val resultFile = File("src/test/assets/very-long-json-lines_scancode-2.2.1.post277.4d68f9377.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -817,7 +817,7 @@ class ScanCodeResultParserTest : WordSpec({
 
         "return false for scan results with other unknown errors" {
             val resultFile = File("src/test/assets/kotlin-annotation-processing-gradle-1.2.21_scancode.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
@@ -835,7 +835,7 @@ class ScanCodeResultParserTest : WordSpec({
 
         "return false for scan results without errors" {
             val resultFile = File("src/test/assets/esprima-2.7.3_scancode-2.2.1.json")
-            val result = parseScanCodeResult(resultFile)
+            val result = parseResultsFile(resultFile)
 
             // The "scanPath" argument should point to the path that was scanned, but as the scanned files are
             // not available here anymore, instead pass the "resultFile" to test the calculation of the package
